Add files via upload
The "Volume Control Using Hand Tracking Module" is a project that leverages OpenCV and a hand-tracking algorithm to control the system's volume based on hand gestures. Using a camera, the module detects the user's hand in real-time, identifies specific gestures or hand positions, and adjusts the volume accordingly. OpenCV, along with libraries like Mediapipe, is used to track hand landmarks and detect movements, such as the distance between the thumb and index finger, which correlates to volume levels. This project demonstrates the application of computer vision and gesture recognition in creating interactive, touchless control systems.